requirements.txt
Flask==0.9
py-bcrypt
-e https://github.com/ohack/python-oliners.git#gg=oliners